requirements.txt
fastapi==0.115.8
pydantic==2.10.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
langchain==0.3.17
langchain-google-genai==2.0.9
oracledb==2.5.1
pydantic-settings==2.9.1
langchain-pinecone==0.2.6
langchain-huggingface==0.2.0
mysql-connector-python==9.3.0 
mariadb==1.1.12